requirements.txt
# requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
boto3==1.28.57
botocore==1.31.57
python-dotenv==0.19.0